# backend-api/docs/openapiSpec.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Watchshop API",
    "version": "1.0.0",
    "description": "Watchshop API",
    "contact": {
      "name": "Thu Quang"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
    "/api/v1/customers/register": {
      "post": {
        "summary": "Create a new customer",
        "description": "Create a new customer",
        "operationId": "createCustomer",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            }
          }
        },
        "tags": [
          "customers"
        ],
        "responses": {
          "201": {
            "description": "A new customer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "description": "The response status",
                      "enum": [
                        "success"
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "customer": {
                          "$ref": "#/components/schemas/Customer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "$ref": "#/components/responses/400BadRequest"
          }
        }
      }
    },
    "/api/v1/customers/login": {
      "post": {
        "summary": "Check for customer",
        "description": "Check for customer",
        "operationId": "CheckCustomer",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CustomerLogin"
              }
            }
          }
        },
        "tags": [
          "customers"
        ],
        "responses": {
          "201": {
            "description": "A new customer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "description": "The response status",
                      "enum": [
                        "success"
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "customer": {
                          "$ref": "#/components/schemas/Customer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "$ref": "#/components/responses/400BadRequest"
          }
        }
      }
    },
    "/api/v1/customers/{id}": {
      "get": {
        "summary": "Get customer by ID",
        "description": "Get customer by ID",
        "operationId": "getCustomerById",
        "parameters": [
          {
            "$ref": "#/components/parameters/customerIdParam"
          }
        ],
        "tags": [
          "customers"
        ],
        "responses": {
          "200": {
            "description": "A customer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "description": "The response status",
                      "enum": [
                        "success"
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "customer": {
                          "$ref": "#/components/schemas/Customer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "$ref": "#/components/responses/400BadRequest"
          }
        }
      },
      "put": {
        "summary": "Update customer by ID",
        "description": "Update customer by ID",
        "operationId": "updateCustomer",
        "parameters": [
          {
            "$ref": "#/components/parameters/customerIdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            }
          }
        },
        "tags": [
          "customers"
        ],
        "responses": {
          "200": {
            "description": "An updated customer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "description": "The response status",
                      "enum": [
                        "success"
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "customer": {
                          "$ref": "#/components/schemas/Customer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "$ref": "#/components/responses/400BadRequest"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "description": "customer name"
          },
          "password": {
            "type": "string",
            "description": "customer password"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "customer email"
          },
          "address": {
            "type": "string",
            "description": "customer address"
          },
          "phone": {
            "type": "string",
            "description": "customer phone number"
          }
        }
      },
      "CustomerLogin": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Customer's email"
          },
          "password": {
            "type": "string",
            "description": "Customer's password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    },
    "responses": {
...
      "400BadRequest": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "message": {
                  "type": "string",
                  "example": "Invalid request parameters"
                }
              }
            }
          }
        }
      },
...
    },
    "parameters": {
      "customerIdParam": {
        "name": "id",
        "description": "Customer ID",
        "required": true,
        "in": "path",
        "schema": {
          "type": "integer"
        }
      },
...
    }
  },
...
}